optional support for webhooks.
webhooks are sent every time the accountant reduces the hour balance of a host.
only when the host has less than 25 hours left.
intended to be used by an automatic system listening on the other side.
to create a host with a webhook, call /create/<image>?notify=https://myurl.com/